Reorganize modules and dedupe cell-mutation/execution-layer code
