+++ Cargo.toml
@@ -30,7 +30,6 @@ sha2 = "0.11"
 base64 = "0.22"
 
 # Local mode dependencies
-tempfile = "3.8"
 dirs = "6.0"
 
 # Connect command dependencies
@@ -55,3 +54,6 @@ ratatui = "0.30"
 crossterm = "0.29"
 syntect = "5.2"
 notify = "8.2"
+
+[dev-dependencies]
+tempfile = "3.8"
