+++ src/commands/cell/add.rs
@@ -1,10 +1,10 @@
 use crate::commands::common::{self, CellType, OutputFormat};
-use crate::notebook;
+use crate::execution::server::ydoc_notebook_ops;
+use crate::notebook::session::{resolve_backend, run_mutation, CellMutator};
 use anyhow::{bail, Context, Result};
 use clap::Parser;
-use nbformat::v4::{Cell, CellId, CellMetadata};
+use nbformat::v4::{Cell, CellId, CellMetadata, Notebook};
 use serde::Serialize;
-use std::collections::HashMap;
 use uuid::Uuid;
 
 #[derive(Parser, Clone)]
@@ -73,7 +73,7 @@ struct AddCellsResult {
     cells: Vec<AddedCellInfo>,
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Clone)]
 struct AddedCellInfo {
     cell_type: String,
     cell_id: String,
@@ -220,347 +220,6 @@ fn parse_source_into_cells(text: &str, default_type: &CellType) -> Vec<ParsedCel
     })
 }
 
-pub fn execute(args: AddCellArgs) -> Result<()> {
-    use crate::execution::types::ExecutionMode;
-    let mode = common::resolve_execution_mode(args.server.clone(), args.token.clone())?;
-
-    match &mode {
-        ExecutionMode::Local | ExecutionMode::RemoteKernel { .. } => execute_file_based(args),
-        ExecutionMode::Remote { .. } => {
-            let runtime = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()?;
-            runtime.block_on(execute_remote(args, mode))
-        }
-    }
-}
-
-/// Remote dispatch: cached Some(false) goes straight to the Contents API;
-/// otherwise try the Y.js realtime path and fall back on the definitive
-/// backend-absent signal. Transient errors propagate so a flaky
-/// collaboration server is never silently downgraded.
-async fn execute_remote(
-    args: AddCellArgs,
-    mode: crate::execution::types::ExecutionMode,
-) -> Result<()> {
-    let cached = common::resolve_ydoc_available(&args.server, &args.token);
-    if cached == Some(false) {
-        return execute_with_contents_api(args, mode).await;
-    }
-    match execute_with_realtime(args.clone(), mode.clone()).await {
-        Err(e) if crate::execution::remote::ydoc::is_yjs_unavailable(&e) => {
-            common::warn_stale_collab_cache(cached);
-            execute_with_contents_api(args, mode).await
-        }
-        result => result,
-    }
-}
-
-async fn execute_with_realtime(
-    args: AddCellArgs,
-    mode: crate::execution::types::ExecutionMode,
-) -> Result<()> {
-    use crate::execution::remote::ydoc_notebook_ops;
-
-    let (server_url, token) = match &mode {
-        crate::execution::types::ExecutionMode::Remote { server_url, token } => {
-            (server_url.clone(), token.clone())
-        }
-        _ => bail!("Expected remote execution mode"),
-    };
-
-    // Normalize notebook path
-    let file_path = common::normalize_notebook_path(&args.file);
-
-    // Compute notebook path relative to server root for Y.js connection
-    let server_root = common::resolve_server_root();
-    let notebook_server_path = common::notebook_path_for_server(&file_path, server_root.as_deref());
-
-    // Read notebook from server (reuse the server-relative path)
-    let notebook = common::read_notebook_remote(&server_url, &token, &notebook_server_path).await?;
-
-    // Parse source content into cells
-    let raw_text = common::parse_source_text(&args.source)?;
-    let parsed_cells = parse_source_into_cells(&raw_text, &args.cell_type);
-
-    // Validate: --id can only be used with a single cell
-    if parsed_cells.len() > 1 && args.id.is_some() {
-        bail!("--id cannot be used when adding multiple cells");
-    }
-
-    // Determine insertion index
-    let insert_index = if let Some(idx) = args.insert_at {
-        if idx < 0 {
-            let abs_idx = idx.unsigned_abs() as usize;
-            if abs_idx > notebook.cells.len() {
-                bail!(
-                    "Negative index {} out of range (notebook has {} cells)",
-                    idx,
-                    notebook.cells.len()
-                );
-            }
-            notebook.cells.len() - abs_idx
-        } else {
-            let pos_idx = idx as usize;
-            if pos_idx > notebook.cells.len() {
-                bail!(
-                    "Index {} out of range (notebook has {} cells)",
-                    idx,
-                    notebook.cells.len()
-                );
-            }
-            pos_idx
-        }
-    } else if let Some(ref after_id) = args.after {
-        let (index, _) = common::find_cell_by_id(&notebook.cells, after_id)?;
-        index + 1
-    } else if let Some(ref before_id) = args.before {
-        let (index, _) = common::find_cell_by_id(&notebook.cells, before_id)?;
-        index
-    } else {
-        notebook.cells.len()
-    };
-
-    // Create all cells
-    let mut new_cells: Vec<Cell> = Vec::new();
-    let mut added_cells: Vec<AddedCellInfo> = Vec::new();
-
-    for (i, parsed) in parsed_cells.into_iter().enumerate() {
-        let cell_id = if let Some(ref id) = args.id {
-            if notebook.cells.iter().any(|c| c.id().as_str() == *id) {
-                bail!("Cell ID '{}' already exists in notebook", id);
-            }
-            CellId::new(id).map_err(|e| anyhow::anyhow!("Invalid cell ID: {}", e))?
-        } else {
-            CellId::from(Uuid::new_v4())
-        };
-
-        let cell_type_str = cell_type_to_str(&parsed.cell_type);
-        let metadata = parsed.metadata.unwrap_or_else(create_empty_metadata);
-        let new_cell = create_cell(parsed.cell_type, cell_id.clone(), metadata, parsed.source);
-
-        added_cells.push(AddedCellInfo {
-            cell_type: cell_type_str.to_string(),
-            cell_id: cell_id.to_string(),
-            index: insert_index + i,
-        });
-
-        new_cells.push(new_cell);
-    }
-
-    // Add cells via Y.js (don't write to file - let JupyterLab handle persistence)
-    ydoc_notebook_ops::ydoc_add_cells(
-        &server_url,
-        &token,
-        &notebook_server_path,
-        &new_cells,
-        insert_index,
-    )
-    .await
-    .context("Error adding cells")?;
-
-    // Output result
-    let format = if args.json {
-        OutputFormat::Json
-    } else {
-        OutputFormat::Text
-    };
-
-    let num_added = added_cells.len();
-    output_results(
-        &file_path,
-        added_cells,
-        notebook.cells.len() + num_added,
-        &format,
-    )?;
-
-    Ok(())
-}
-
-async fn execute_with_contents_api(
-    args: AddCellArgs,
-    mode: crate::execution::types::ExecutionMode,
-) -> Result<()> {
-    let file = args.file.clone();
-    let (file_path, added_cells, total_cells) =
-        common::with_contents_api(&file, &mode, |notebook, file_path| {
-            let raw_text = common::parse_source_text(&args.source)?;
-            let parsed_cells = parse_source_into_cells(&raw_text, &args.cell_type);
-
-            if parsed_cells.len() > 1 && args.id.is_some() {
-                bail!("--id cannot be used when adding multiple cells");
-            }
-
-            let insert_index = if let Some(idx) = args.insert_at {
-                if idx < 0 {
-                    let abs_idx = idx.unsigned_abs() as usize;
-                    if abs_idx > notebook.cells.len() {
-                        bail!(
-                            "Negative index {} out of range (notebook has {} cells)",
-                            idx,
-                            notebook.cells.len()
-                        );
-                    }
-                    notebook.cells.len() - abs_idx
-                } else {
-                    let pos_idx = idx as usize;
-                    if pos_idx > notebook.cells.len() {
-                        bail!(
-                            "Index {} out of range (notebook has {} cells)",
-                            idx,
-                            notebook.cells.len()
-                        );
-                    }
-                    pos_idx
-                }
-            } else if let Some(ref after_id) = args.after {
-                let (index, _) = common::find_cell_by_id(&notebook.cells, after_id)?;
-                index + 1
-            } else if let Some(ref before_id) = args.before {
-                let (index, _) = common::find_cell_by_id(&notebook.cells, before_id)?;
-                index
-            } else {
-                notebook.cells.len()
-            };
-
-            let mut added_cells: Vec<AddedCellInfo> = Vec::new();
-
-            for (i, parsed) in parsed_cells.into_iter().enumerate() {
-                let cell_id = if let Some(ref id) = args.id {
-                    if notebook.cells.iter().any(|c| c.id().as_str() == *id) {
-                        bail!("Cell ID '{}' already exists in notebook", id);
-                    }
-                    CellId::new(id).map_err(|e| anyhow::anyhow!("Invalid cell ID: {}", e))?
-                } else {
-                    CellId::from(Uuid::new_v4())
-                };
-
-                let cell_type_str = cell_type_to_str(&parsed.cell_type);
-                let metadata = parsed.metadata.unwrap_or_else(create_empty_metadata);
-                let new_cell =
-                    create_cell(parsed.cell_type, cell_id.clone(), metadata, parsed.source);
-
-                let actual_index = insert_index + i;
-                notebook.cells.insert(actual_index, new_cell);
-
-                added_cells.push(AddedCellInfo {
-                    cell_type: cell_type_str.to_string(),
-                    cell_id: cell_id.to_string(),
-                    index: actual_index,
-                });
-            }
-
-            Ok((file_path.to_string(), added_cells, notebook.cells.len()))
-        })
-        .await?;
-
-    let format = if args.json {
-        OutputFormat::Json
-    } else {
-        OutputFormat::Text
-    };
-    output_results(&file_path, added_cells, total_cells, &format)
-}
-
-fn execute_file_based(args: AddCellArgs) -> Result<()> {
-    // Normalize notebook path
-    let file_path = common::normalize_notebook_path(&args.file);
-
-    // Read notebook
-    let mut notebook = notebook::read_notebook(&file_path).context("Failed to read notebook")?;
-
-    // Parse source content into cells
-    let raw_text = common::parse_source_text(&args.source)?;
-    let parsed_cells = parse_source_into_cells(&raw_text, &args.cell_type);
-
-    // Validate: --id can only be used with a single cell
-    if parsed_cells.len() > 1 && args.id.is_some() {
-        bail!("--id cannot be used when adding multiple cells");
-    }
-
-    // Determine insertion index
-    let insert_index = if let Some(idx) = args.insert_at {
-        if idx < 0 {
-            let abs_idx = idx.unsigned_abs() as usize;
-            if abs_idx > notebook.cells.len() {
-                bail!(
-                    "Negative index {} out of range (notebook has {} cells)",
-                    idx,
-                    notebook.cells.len()
-                );
-            }
-            notebook.cells.len() - abs_idx
-        } else {
-            let pos_idx = idx as usize;
-            if pos_idx > notebook.cells.len() {
-                bail!(
-                    "Index {} out of range (notebook has {} cells)",
-                    idx,
-                    notebook.cells.len()
-                );
-            }
-            pos_idx
-        }
-    } else if let Some(ref after_id) = args.after {
-        let (index, _) = common::find_cell_by_id(&notebook.cells, after_id)?;
-        index + 1
-    } else if let Some(ref before_id) = args.before {
-        let (index, _) = common::find_cell_by_id(&notebook.cells, before_id)?;
-        index
-    } else {
-        notebook.cells.len()
-    };
-
-    // Create and insert cells
-    let mut added_cells: Vec<AddedCellInfo> = Vec::new();
-
-    for (i, parsed) in parsed_cells.into_iter().enumerate() {
-        let cell_id = if let Some(ref id) = args.id {
-            if notebook.cells.iter().any(|c| c.id().as_str() == *id) {
-                bail!("Cell ID '{}' already exists in notebook", id);
-            }
-            CellId::new(id).map_err(|e| anyhow::anyhow!("Invalid cell ID: {}", e))?
-        } else {
-            CellId::from(Uuid::new_v4())
-        };
-
-        let cell_type_str = cell_type_to_str(&parsed.cell_type);
-        let metadata = parsed.metadata.unwrap_or_else(create_empty_metadata);
-        let new_cell = create_cell(parsed.cell_type, cell_id.clone(), metadata, parsed.source);
-
-        let actual_index = insert_index + i;
-        notebook.cells.insert(actual_index, new_cell);
-
-        added_cells.push(AddedCellInfo {
-            cell_type: cell_type_str.to_string(),
-            cell_id: cell_id.to_string(),
-            index: actual_index,
-        });
-    }
-
-    // Write notebook atomically
-    notebook::write_notebook_atomic(&file_path, &notebook).context("Failed to write notebook")?;
-
-    // Output result
-    let format = if args.json {
-        OutputFormat::Json
-    } else {
-        OutputFormat::Text
-    };
-
-    output_results(&file_path, added_cells, notebook.cells.len(), &format)?;
-
-    Ok(())
-}
-
-fn cell_type_to_str(ct: &CellType) -> &'static str {
-    match ct {
-        CellType::Code => "code",
-        CellType::Markdown => "markdown",
-        CellType::Raw => "raw",
-    }
-}
-
 fn create_cell(
     cell_type: CellType,
     id: CellId,
@@ -589,20 +248,158 @@ fn create_cell(
     }
 }
 
-fn create_empty_metadata() -> nbformat::v4::CellMetadata {
-    nbformat::v4::CellMetadata {
-        id: None,
-        collapsed: None,
-        scrolled: None,
-        deletable: None,
-        editable: None,
-        format: None,
-        name: None,
-        tags: None,
-        jupyter: None,
-        execution: None,
-        additional: HashMap::new(),
+/// Resolve the index at which new cells should be inserted, from
+/// `--insert-at`/`--after`/`--before` (mutually exclusive; default: append).
+fn resolve_insert_index(args: &AddCellArgs, cells: &[Cell]) -> Result<usize> {
+    if let Some(idx) = args.insert_at {
+        if idx < 0 {
+            let abs_idx = idx.unsigned_abs() as usize;
+            if abs_idx > cells.len() {
+                bail!(
+                    "Negative index {} out of range (notebook has {} cells)",
+                    idx,
+                    cells.len()
+                );
+            }
+            Ok(cells.len() - abs_idx)
+        } else {
+            let pos_idx = idx as usize;
+            if pos_idx > cells.len() {
+                bail!(
+                    "Index {} out of range (notebook has {} cells)",
+                    idx,
+                    cells.len()
+                );
+            }
+            Ok(pos_idx)
+        }
+    } else if let Some(ref after_id) = args.after {
+        let (index, _) = common::find_cell_by_id(cells, after_id)?;
+        Ok(index + 1)
+    } else if let Some(ref before_id) = args.before {
+        let (index, _) = common::find_cell_by_id(cells, before_id)?;
+        Ok(index)
+    } else {
+        Ok(cells.len())
     }
+}
+
+struct AddCellMutator<'a> {
+    args: &'a AddCellArgs,
+}
+
+#[async_trait::async_trait]
+impl CellMutator for AddCellMutator<'_> {
+    type Output = (Vec<AddedCellInfo>, usize);
+
+    fn mutate_notebook(&self, notebook: &mut Notebook, _file_path: &str) -> Result<Self::Output> {
+        let raw_text = common::parse_source_text(&self.args.source)?;
+        let parsed_cells = parse_source_into_cells(&raw_text, &self.args.cell_type);
+
+        if parsed_cells.len() > 1 && self.args.id.is_some() {
+            bail!("--id cannot be used when adding multiple cells");
+        }
+
+        let insert_index = resolve_insert_index(self.args, &notebook.cells)?;
+        let mut added_cells: Vec<AddedCellInfo> = Vec::new();
+
+        for (i, parsed) in parsed_cells.into_iter().enumerate() {
+            let cell_id = new_cell_id(self.args, &notebook.cells)?;
+            let cell_type_str = common::cell_type_enum_str(&parsed.cell_type);
+            let metadata = parsed.metadata.unwrap_or_else(common::empty_cell_metadata);
+            let new_cell = create_cell(parsed.cell_type, cell_id.clone(), metadata, parsed.source);
+
+            let actual_index = insert_index + i;
+            notebook.cells.insert(actual_index, new_cell);
+
+            added_cells.push(AddedCellInfo {
+                cell_type: cell_type_str.to_string(),
+                cell_id: cell_id.to_string(),
+                index: actual_index,
+            });
+        }
+
+        Ok((added_cells, notebook.cells.len()))
+    }
+
+    async fn mutate_realtime(
+        &self,
+        server_url: &str,
+        token: &str,
+        server_path: &str,
+        _file_path: &str,
+    ) -> Result<Self::Output> {
+        let notebook =
+            crate::notebook::remote::read_notebook_remote(server_url, token, server_path).await?;
+
+        let raw_text = common::parse_source_text(&self.args.source)?;
+        let parsed_cells = parse_source_into_cells(&raw_text, &self.args.cell_type);
+
+        if parsed_cells.len() > 1 && self.args.id.is_some() {
+            bail!("--id cannot be used when adding multiple cells");
+        }
+
+        let insert_index = resolve_insert_index(self.args, &notebook.cells)?;
+
+        let mut new_cells: Vec<Cell> = Vec::new();
+        let mut added_cells: Vec<AddedCellInfo> = Vec::new();
+
+        for (i, parsed) in parsed_cells.into_iter().enumerate() {
+            let cell_id = new_cell_id(self.args, &notebook.cells)?;
+            let cell_type_str = common::cell_type_enum_str(&parsed.cell_type);
+            let metadata = parsed.metadata.unwrap_or_else(common::empty_cell_metadata);
+            let new_cell = create_cell(parsed.cell_type, cell_id.clone(), metadata, parsed.source);
+
+            added_cells.push(AddedCellInfo {
+                cell_type: cell_type_str.to_string(),
+                cell_id: cell_id.to_string(),
+                index: insert_index + i,
+            });
+
+            new_cells.push(new_cell);
+        }
+
+        let num_added = added_cells.len();
+
+        // Add cells via Y.js (don't write to file - let JupyterLab handle persistence)
+        ydoc_notebook_ops::ydoc_add_cells(server_url, token, server_path, &new_cells, insert_index)
+            .await
+            .context("Error adding cells")?;
+
+        Ok((added_cells, notebook.cells.len() + num_added))
+    }
+}
+
+/// Generate (or validate a user-supplied) cell ID, checking for collisions
+/// against the existing cells.
+fn new_cell_id(args: &AddCellArgs, cells: &[Cell]) -> Result<CellId> {
+    if let Some(ref id) = args.id {
+        if cells.iter().any(|c| c.id().as_str() == *id) {
+            bail!("Cell ID '{}' already exists in notebook", id);
+        }
+        CellId::new(id).map_err(|e| anyhow::anyhow!("Invalid cell ID: {}", e))
+    } else {
+        Ok(CellId::from(Uuid::new_v4()))
+    }
+}
+
+pub fn execute(args: AddCellArgs) -> Result<()> {
+    let backend = resolve_backend(&args.file, args.server.clone(), args.token.clone())?;
+    let mutator = AddCellMutator { args: &args };
+
+    let runtime = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()?;
+    let (file_path, (added_cells, total_cells)) =
+        runtime.block_on(run_mutation(backend, &mutator))?;
+
+    let format = if args.json {
+        OutputFormat::Json
+    } else {
+        OutputFormat::Text
+    };
+
+    output_results(&file_path, added_cells, total_cells, &format)
 }
 
 fn output_results(
@@ -633,39 +430,27 @@ fn output_results(
 }
 
 fn output_result(result: &AddCellResult, format: &OutputFormat) -> Result<()> {
-    match format {
-        OutputFormat::Json => {
-            println!("{}", serde_json::to_string_pretty(&result)?);
-        }
-        OutputFormat::Text | OutputFormat::Markdown => {
-            println!("Added {} cell to: {}", result.cell_type, result.file);
-            println!("Cell ID: {}", result.cell_id);
-            println!(
-                "Index: {} (total: {} cells)",
-                result.index, result.total_cells
-            );
-        }
-    }
-    Ok(())
+    common::print_result(result, format, |result| {
+        println!("Added {} cell to: {}", result.cell_type, result.file);
+        println!("Cell ID: {}", result.cell_id);
+        println!(
+            "Index: {} (total: {} cells)",
+            result.index, result.total_cells
+        );
+    })
 }
 
 fn output_multi_result(result: &AddCellsResult, format: &OutputFormat) -> Result<()> {
-    match format {
-        OutputFormat::Json => {
-            println!("{}", serde_json::to_string_pretty(&result)?);
+    common::print_result(result, format, |result| {
+        println!("Added {} cells to: {}", result.cells_added, result.file);
+        for cell in &result.cells {
+            println!(
+                "  Cell ID: {} (type: {}, index: {})",
+                cell.cell_id, cell.cell_type, cell.index
+            );
         }
-        OutputFormat::Text | OutputFormat::Markdown => {
-            println!("Added {} cells to: {}", result.cells_added, result.file);
-            for cell in &result.cells {
-                println!(
-                    "  Cell ID: {} (type: {}, index: {})",
-                    cell.cell_id, cell.cell_type, cell.index
-                );
-            }
-            println!("Total cells: {}", result.total_cells);
-        }
-    }
-    Ok(())
+        println!("Total cells: {}", result.total_cells);
+    })
 }
 
 #[cfg(test)]
