+++ src/commands/connection/mod.rs
@@ -1,0 +1,3 @@
+pub mod connect;
+pub mod disconnect;
+pub mod status;
