+++ src/commands/mod.rs
@@ -1,17 +1,10 @@
-pub mod add_cell;
-pub mod clean_output_dirs;
-pub mod clear_outputs;
+pub mod cell;
 pub mod common;
-pub mod connect;
+pub mod connection;
 pub mod create_notebook;
-pub mod delete_cell;
-pub mod disconnect;
-pub mod env_manager;
 pub mod execute_notebook;
-pub mod markdown_renderer;
+pub mod output;
 pub mod read;
 pub mod search;
 pub mod self_cmd;
-pub mod status;
-pub mod update_cell;
 pub mod view;
