+++ src/commands/output/mod.rs
@@ -1,0 +1,3 @@
+pub mod clean;
+pub mod clear;
+pub mod markdown_renderer;
