+++ src/execution/local/mod.rs
@@ -2,6 +2,5 @@
 
 pub mod discovery;
 pub mod executor;
-pub mod manager;
 
 pub use executor::LocalExecutor;
