+++ src/execution/types.rs
@@ -1,8 +1,7 @@
-use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use std::time::Duration;
 
-use crate::commands::env_manager::EnvConfig;
+use crate::execution::env::EnvConfig;
 
 /// Execution mode: local (direct kernel), remote (Jupyter server), or remote-kernel (gateway)
 #[derive(Debug, Clone, PartialEq)]
@@ -115,77 +114,6 @@ impl ExecutionResult {
             outputs,
             execution_count,
             error: Some(error),
-        }
-    }
-}
-
-/// Output from a single execution message
-#[allow(dead_code)]
-#[derive(Debug)]
-pub enum MessageOutput {
-    /// stdout/stderr stream
-    Stream { name: String, text: String },
-
-    /// Display data (images, HTML, etc.)
-    DisplayData {
-        data: serde_json::Value,
-        metadata: serde_json::Value,
-    },
-
-    /// Execute result (output of cell)
-    ExecuteResult {
-        data: serde_json::Value,
-        metadata: serde_json::Value,
-        execution_count: i64,
-    },
-
-    /// Error/exception
-    Error {
-        ename: String,
-        evalue: String,
-        traceback: Vec<String>,
-    },
-}
-
-impl MessageOutput {
-    /// Convert to nbformat Output
-    #[allow(dead_code)]
-    pub fn to_nbformat_output(&self) -> Result<nbformat::v4::Output> {
-        match self {
-            MessageOutput::Stream { name, text } => Ok(nbformat::v4::Output::Stream {
-                name: name.clone(),
-                text: nbformat::v4::MultilineString(text.clone()),
-            }),
-            MessageOutput::DisplayData { data, metadata } => {
-                let json = serde_json::json!({
-                    "output_type": "display_data",
-                    "data": data,
-                    "metadata": metadata
-                });
-                Ok(serde_json::from_value(json)?)
-            }
-            MessageOutput::ExecuteResult {
-                data,
-                metadata,
-                execution_count,
-            } => {
-                let json = serde_json::json!({
-                    "output_type": "execute_result",
-                    "execution_count": execution_count,
-                    "data": data,
-                    "metadata": metadata
-                });
-                Ok(serde_json::from_value(json)?)
-            }
-            MessageOutput::Error {
-                ename,
-                evalue,
-                traceback,
-            } => Ok(nbformat::v4::Output::Error(nbformat::v4::ErrorOutput {
-                ename: ename.clone(),
-                evalue: evalue.clone(),
-                traceback: traceback.clone(),
-            })),
         }
     }
 }
